ol, tl, hom, and m6
Added copying offset library, tool library, hom file IF custom home setting is set, and mg IF custom m6 setting is set. Also fixed an issue with wizardsetting.xml where if there were a duplicate setting it would set both to the lower of the two, which is incorrect. Now both are set to the higher one, which is correct.