Cleanup
Fixed a "bug" where ATCWritten wasn't being set to True when the wizard writes but the ATC menu isn't opened. Manually writing ATCWritten to True to get around this, when custom M6 is selected. Also added the MPG parameters to the 5.20+ params. Going to generate an EXE off of this and barring any serious problems will probably call this "done."